Add files via upload
I wrote a small program to simulate a traffic light and crosswalk signal set.  The setup begins with the traffic light green, and waits for a push button input from the pedestrian.  The traffic light goes to yellow, then red, after which the "walk" light illuminates for a time.  The "do not walk" light then illuminates and the traffic light returns to green.